define core multi_touch class
